WIP - Initial implementation of .

+++ src/layup/utilities/file_io/file_output.py
@@ -1,7 +1,7 @@
 """Utility functions for writing data from our internal representation of numpy structured arrays to output files."""
 
 import logging
-
+import os
 import pandas as pd
 
 
@@ -16,7 +16,12 @@ def write_csv(data, filepath):
         The path to the file to write.
     """
     df = pd.DataFrame(data)
-    df.to_csv(filepath, index=False)
+    if os.path.exists(filepath):
+        with open(filepath, "a") as f:
+            df.to_csv(f, index=False)
+    else:
+        df.to_csv(filepath, index=False)
+
     logging.info(f"Data written to {filepath}")
 
 
@@ -33,5 +38,9 @@ def write_hdf5(data, filepath, key="data"):
         The key to use in the HDF5 file.
     """
     df = pd.DataFrame(data)
-    df.to_hdf(filepath, key, mode="w", format="table")
+
+    store = pd.HDFStore(filepath)
+    store.append(key, df, format="t", data_columns=True)
+    store.close()
+
     logging.info(f"Data written to {filepath}")
